+++ engine/src/main/kotlin/engine/patterns/OperatorPattern.kt
@@ -32,8 +32,6 @@ import engine.operators.SetOperators
 import engine.operators.SolutionOperator
 import engine.operators.SubtractEquationsOperator
 import engine.operators.UnaryExpressionOperator
-import engine.operators.UnitExpressionOperator
-import engine.operators.UnitType
 import engine.operators.VariableListOperator
 
 /**
@@ -133,5 +131,3 @@ fun closedOpenIntervalOf(lhs: Pattern, rhs: Pattern) =
 
 fun closedIntervalOf(lhs: Pattern, rhs: Pattern) =
     OperatorPattern(IntervalOperator(closedLeft = true, closedRight = true), listOf(lhs, rhs))
-
-fun degreeOf(value: Pattern) = OperatorPattern(UnitExpressionOperator(unit = UnitType.Degree), listOf(value))